+++ packages/hooks/src/components/use-tree/index.ts
@@ -1,5 +1,11 @@
 export { default, default as useTree } from './use-tree';
-export type { BaseTreeProps, UpdateFunc, TreePathType, TreeModeType } from './use-tree.type';
+export type {
+  BaseTreeProps,
+  UpdateFunc,
+  TreePathType,
+  TreeModeType,
+  FlatNodeType,
+} from './use-tree.type';
 
 export { default as useTreeNode } from './use-tree-node';
 export type { BaseTreeNodeProps, UpdateType } from './use-tree-node.type';
